refactor: move test files from test folder to src folder
