+++ eCommerce-Application/src/api/client/client.test.ts
@@ -1,5 +1,5 @@
 import { HttpMiddlewareOptions } from '@commercetools/ts-client';
-import { client } from '../../../src/api/client/client';
+import { client } from './client';
 import { test, expect } from 'vitest';
 
 const httpMiddleware: HttpMiddlewareOptions = {
